// src/Juniper.TypeScript/@juniper-lib/dom/tags.d.ts
export interface ErsatzElement<T extends Element = Element> {
    element: T;
}